chore: move http_error to customErrors package
