+++ internal/customErrors/http_error.go
@@ -1,4 +1,4 @@
-package internal
+package customErrors
 
 type HTTPError struct {
 	Status int
